use idempotency key to prevent duplicate payments

+++ models/billing.go
@@ -14,9 +14,9 @@ type User struct {
 
 // ServicePlan represents a service plan with pricing and limits
 type ServicePlan struct {
-	BaseCosts       float64
-	MinutesPerMonth float64
-	RecordingSpace  float64
+	BaseCosts       int64
+	MinutesPerMonth int64
+	RecordingSpace  int64
 	Fax             int
 	PayAsYouGo      bool
 }
@@ -24,12 +24,12 @@ type ServicePlan struct {
 // BillingInfo contains billing information for a workspace
 type BillingInfo struct {
 	InvoiceDue            string
-	RemainingBalanceCents float64
+	RemainingBalanceCents int64
 }
 
 // BaseCosts contains base cost rates for various services
 type BaseCosts struct {
-	RecordingsPerByte float64
+	RecordingsPerByte int64
 	FaxPerUsed        float64
 }
